+++ bundles/framework/rpc/instance.js
@@ -260,6 +260,12 @@ Oskari.clazz.define(
             },
             resetState : function() {
                 this.sandbox.resetState();
+            },
+            getCurrentState : function() {
+                return this.sandbox.getCurrentState();
+            },
+            useState : function(state) {
+                this.sandbox.useState(state);
             }
         },
 
@@ -276,14 +282,14 @@ Oskari.clazz.define(
             var me = this,
                 funcs = this._allowedFunctions;
             var bindFunction = function(name) {
-                channel.bind(name, function (trans) {
+                channel.bind(name, function (trans, params) {
                     if (!me._domainMatch(trans.origin)) {
                         throw {
                             error: 'invalid_origin',
                             message: 'Invalid origin: ' + trans.origin
                         };
                     }
-                    return me._availableFunctions[name].apply(me);
+                    return me._availableFunctions[name].apply(me, params);
                 });
             }